some controllers

+++ Backend/Model/Hotel/Hotel-model.go
@@ -1,7 +1,7 @@
 package Model
 
 type Hotel struct {
-	HotelID uint   `gorm:"primary_key;autoIncrement:true;unique"`
+	HotelID int64  `gorm:"primary_key;autoIncrement:true;unique"`
 	Name    string `gorm:"type:varchar(255);not null"`
 	Rooms   int64  `gorm:"type:integer(255);not null"`
 }
